refine routes and file structure
