--- client/landing/jetpack-cloud/sections/settings/settings-flow/style.scss
@@ -1,5 +1,0 @@
-.settings-flow__notice {
-    padding: 21px 24px 24px 27px;
-    margin: 24px 0;
-    background: var( --studio-yellow-0 );
-}
